move deisotoping
