+++ src/openms/includes.cmake
@@ -45,6 +45,7 @@ include(source/PROCESSING/BASELINE/sources.cmake)
 include(source/PROCESSING/FILTERING/sources.cmake)
 include(source/PROCESSING/RESAMPLING/sources.cmake)
 include(source/PROCESSING/DATAREDUCTION/sources.cmake)
+include(source/PROCESSING/DEISOTOPING/sources.cmake)
 include(source/PROCESSING/CALIBRATION/sources.cmake)
 include(source/PROCESSING/SMOOTHING/sources.cmake)
 include(source/PROCESSING/NOISEESTIMATION/sources.cmake)
@@ -112,6 +113,7 @@ include(include/OpenMS/COMPARISON/sources.cmake)
 include(include/OpenMS/PROCESSING/BASELINE/sources.cmake)
 include(include/OpenMS/PROCESSING/CALIBRATION/sources.cmake)
 include(include/OpenMS/PROCESSING/DATAREDUCTION/sources.cmake)
+include(include/OpenMS/PROCESSING/DEISOTOPING/sources.cmake)
 include(include/OpenMS/PROCESSING/ID/sources.cmake)
 include(include/OpenMS/PROCESSING/NOISEESTIMATION/sources.cmake)
 include(include/OpenMS/PROCESSING/SMOOTHING/sources.cmake)
